requirements.txt
flask
Flask-WTF
python-dotenv
flask-login
bcrypt
apsw